+++ lib/superform/rails/components/textarea.rb
@@ -6,6 +6,15 @@ module Superform
           content ||= Proc.new { dom.value }
           textarea(**attributes, &content)
         end
+
+        def field_attributes
+          attrs = {
+            id: dom.id,
+            name: dom.name
+          }
+          attrs[:readonly] = true if field.read_only?
+          attrs
+        end
       end
     end
   end